+++ src/ballistica/ballistica.cc
@@ -21,7 +21,7 @@
 namespace ballistica {
 
 // These are set automatically via script; don't change here.
-const int kAppBuildNumber = 20300;
+const int kAppBuildNumber = 20303;
 const char* kAppVersion = "1.6.0";
 
 // Our standalone globals.